Initial version of Postgres 12 support
Please note that this is still work in progress, and in particular the
fingerprinting logic will still change.

+++ src/postgres/src_backend_nodes_value.c
@@ -12,7 +12,7 @@
  *	  implementation of Value nodes
  *
  *
- * Copyright (c) 2003-2017, PostgreSQL Global Development Group
+ * Copyright (c) 2003-2019, PostgreSQL Global Development Group
  *
  *
  * IDENTIFICATION
@@ -28,7 +28,7 @@
  *	makeInteger
  */
 Value *
-makeInteger(long i)
+makeInteger(int i)
 {
 	Value	   *v = makeNode(Value);
